Illumos 7090 overflows stack with zio_{re}execute calls
Symptom: memory corruption panic under heavy write load in
the presence of metadata read checksum errors.

Typically this happens in a taskq task that destroys an
arc_buf.

This symptom appears even more reliably when inlining is disabled at
compile time.

ZoL's zio_execute() was originally decorated with
__attribute__((always_inline)) which was not used for xnu.

Follow ZoL logic wrt stack overflow in light of 7090 stack
overflow, (re)applying it to zio_execute(), and also to
zio_reexecute() and zio_taskq_dispatch().

Additionally, make zfs_vdev_queue_depth_pct and
zio_dva_throttle_enabled sysctl tunables.

Generate an ecksum ereport in zio_done(); this was
previously commented out.

    enable ECKSUM ereport in zio_done()

    <lundman> ok OPPLE it is
    [08:43:28]

+++ include/sys/vdev_impl.h
@@ -57,7 +57,7 @@ typedef struct vdev_queue vdev_queue_t;
 typedef struct vdev_cache vdev_cache_t;
 typedef struct vdev_cache_entry vdev_cache_entry_t;
 
-extern int zfs_vdev_queue_depth_pct;
+extern uint64_t zfs_vdev_queue_depth_pct;
 extern uint32_t zfs_vdev_async_write_max_active;
 
 /*
